+++ app/assets/javascripts/libs/utils.js
@@ -8,9 +8,10 @@ import type { Vector3, Vector4, Vector6, BoundingBoxType } from "oxalis/constant
 import Maybe from "data.maybe";
 import window, { document, location } from "libs/window";
 import pako from "pako";
+import naturalSort from "javascript-natural-sort";
 import type { APIUserType } from "admin/api_flow_types";
 
-type Comparator<T> = (T, T) => -1 | 0 | 1;
+export type Comparator<T> = (T, T) => -1 | 0 | 1;
 type UrlParamsType = { [key: string]: string | boolean };
 
 function swap(arr, a, b) {
@@ -21,6 +22,8 @@ function swap(arr, a, b) {
     arr[a] = tmp;
   }
 }
+
+naturalSort.insensitive = true;
 
 function getRecursiveValues(obj: Object | Array<*> | string): Array<*> {
   return _.flattenDeep(getRecursiveValuesUnflat(obj));
@@ -34,6 +37,14 @@ function getRecursiveValuesUnflat(obj: Object | Array<*> | string): Array<*> {
   } else {
     return [obj];
   }
+}
+
+function cheapSort<T: string | number>(valueA: T, valueB: T): -1 | 0 | 1 {
+  // $FlowFixMe It is not possible to express that valueA and valueB have the very same type
+  if (valueA < valueB) return -1;
+  // $FlowFixMe It is not possible to express that valueA and valueB have the very same type
+  if (valueA > valueB) return 1;
+  return 0;
 }
 
 export function enforce<A, B>(fn: A => B): (?A) => B {
@@ -147,13 +158,7 @@ const Utils = {
       if (typeof valueA !== "number" || typeof valueB !== "number") {
         return 0;
       }
-      if (valueA < valueB) {
-        return -1;
-      }
-      if (valueA > valueB) {
-        return 1;
-      }
-      return 0;
+      return cheapSort(valueA, valueB);
     };
   },
 
@@ -161,21 +166,19 @@ const Utils = {
     collectionForTypeInference: Array<T>, // this parameter is only used let flow infer the used type
     selector: $Keys<T> | (T => string),
     isSortedAscending: boolean = true,
-  ): (T, T) => number {
-    const sortingOrder = isSortedAscending ? 1 : -1;
-
-    return (a: T, b: T): number => {
+    sortNatural: boolean = true,
+  ): Comparator<T> {
+    return (a: T, b: T) => {
+      if (!isSortedAscending) {
+        [a, b] = [b, a];
+      }
       const valueA = typeof selector === "function" ? selector(a) : a[selector];
       const valueB = typeof selector === "function" ? selector(b) : b[selector];
       if (typeof valueA !== "string" || typeof valueB !== "string") {
         return 0;
       }
-      return (
-        valueA.localeCompare(valueB, "en", {
-          numeric: true,
-          usage: "search",
-        }) * sortingOrder
-      );
+      // localeCompare is really slow, therefore, we use the naturalSort lib and a cheap sorting otherwise
+      return sortNatural ? naturalSort(valueA, valueB) : cheapSort(valueA, valueB);
     };
   },